Moved field variable signup into methods
The signup field variable induces extreme side-effects because all of
the methods are static. As a result, a SignUpPage object is never
created. If you try to run a parameterized test, everything will fail
aside from the first test because the Driver.Instance in the 2nd test
refers to the killed Driver in the 1st test. Since the Driver does not
exist (it was killed after the 1st test), all future tests will fail.